+++ CHANGELOG.md
@@ -8,6 +8,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 
 ### [Unreleased]
 - ListBackups: ported to Rust
+- Cardano: allow transactions with a zero TTL value
 
 ### 9.8.0 [released 2021-10-21]
 - Multi edition: add Cardano support.